+++ runtime/lib/objects/lambda.c
@@ -92,7 +92,7 @@ static CL_Object* cl_lambda_call_array(CL_Object* obj, CL_FUNC_PARAMS) {
 }
 
 // The function is called by an ordinary procedure call
-CL_Object* cl_lambda_call(CL_Object* obj, unsigned int count, ...) {
+CL_Object* cl_lambda_call(CL_Object* obj, size_t count, ...) {
     CL_CHECK_FUNC_ARG_TYPE(cl_get_obj_type(obj), LAMBDA);
 
     va_list args;
@@ -108,18 +108,18 @@ CL_Object* cl_lambda_call(CL_Object* obj, unsigned int count, ...) {
 }
 
 // The function is called by (apply ...)
-CL_Object* cl_lambda_call_list(CL_Object* obj, unsigned int count, ...) {
+CL_Object* cl_lambda_call_list(CL_Object* obj, size_t count, ...) {
     CL_CHECK_FUNC_ARG_TYPE(cl_get_obj_type(obj), LAMBDA);
 
     va_list args;
     va_start(args, count);
     CL_Object* tmp[count];
-    for (unsigned int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         tmp[i] = va_arg(args, CL_Object*);
     }
     va_end(args);
 
-    unsigned int scalar_args_count = count - 1;
+    size_t scalar_args_count = count - 1;
     CL_Object* list_arg = tmp[scalar_args_count];
 
     if (!cl_is_list_internal(list_arg)) {
@@ -128,15 +128,15 @@ CL_Object* cl_lambda_call_list(CL_Object* obj, unsigned int count, ...) {
     }
 
 
-    unsigned int list_arg_length = cl_list_length_internal(list_arg);
-    unsigned int obj_args_count = scalar_args_count + list_arg_length;
+    size_t list_arg_length = cl_list_length_internal(list_arg);
+    size_t obj_args_count = scalar_args_count + list_arg_length;
     CL_Object* obj_args[obj_args_count];
 
-    for (unsigned int i = 0; i < scalar_args_count; i++) {
+    for (size_t i = 0; i < scalar_args_count; i++) {
         obj_args[i] = tmp[i];
     }
 
-    unsigned int curr_list_pos = 0;
+    size_t curr_list_pos = 0;
     while (cl_get_obj_type(list_arg) != EMPTY_LIST) {
         obj_args[scalar_args_count + curr_list_pos++] = cl_get_pair_left_internal(list_arg);
         list_arg = cl_get_pair_right_internal(list_arg);
